Updated StopsActivity to use StopsFragment
