+++ app/src/main/java/edu/mit/mitmobile2/shuttles/ShuttleStopViewPagerFragment.java
@@ -1,9 +1,9 @@
 package edu.mit.mitmobile2.shuttles;
 
+import android.app.Fragment;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -24,7 +24,7 @@ import edu.mit.mitmobile2.shuttles.model.MITShuttlePrediction;
 import edu.mit.mitmobile2.shuttles.model.MITShuttleRoute;
 import edu.mit.mitmobile2.shuttles.model.MITShuttleStopWrapper;
 
-public class ShuttleStopViewPagerFragment extends Fragment{
+public class ShuttleStopViewPagerFragment extends Fragment {
 
     @InjectView(R.id.stop_prediction_adapter_view)
     AdapterView predictionAdapterView;
