+++ app/src/main/res/layout/activity_with_frame.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/route_screen_frame"
+    android:id="@+id/fragment_frame"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent" />
